feat: add constant applicant check
