+++ frontend/components/kanban/ColumnWithBackButton.component.tsx
@@ -3,7 +3,7 @@
 import { KanbanColumnData } from "@/src/stores/kanban/Kanban.atoms";
 import { useAtom } from "jotai";
 import { FC } from "react";
-import KanbanCardComponent from "./Card.component";
+import KanbanCardComponent from "./card/Card.component";
 import KanbanDetailBackButton from "./detail/BackButton.component";
 import { useQuery } from "@tanstack/react-query";
 import { getAllKanbanData } from "@/src/apis/kanban/kanban";
