+++ frontend/components/kanban/card/index.tsx
@@ -1,0 +1,7 @@
+import KanbanCardApplicant from "./CardApplicant.component";
+import KanbanCardInvisible from "./CardInvisible.component";
+
+export const KanbanCard = {
+  Invisible: KanbanCardInvisible,
+  Applicant: KanbanCardApplicant,
+};
